add script to run tests, refactor rtokent tests into new file.
